+++ src/libCrypto/EthCrypto.h
@@ -38,20 +38,4 @@ std::string ToUncompressedPubKey(const std::string& pubKey);
 
 bytes recoverECDSAPubSig(std::string const& message, int chain_id);
 
-struct EthFields {
-  uint32_t version{};
-  uint64_t nonce{};  // counter: the number of tx from m_fromAddr
-  bytes toAddr;
-  uint128_t amount;
-  uint128_t gasPrice;
-  uint64_t gasLimit{};
-  bytes code;
-  bytes data;
-  bytes signature;
-};
-
-EthFields parseRawTxFields(std::string const& message);
-
-PubKey toPubKey(bytes const& key);
-
 #endif  // ZILLIQA_SRC_LIBCRYPTO_ETHCRYPTO_H_
